Various dashboard improvements
The primary change in this PR has to do with error handling. We added an alert to each of our main dashboards which is shown when errors are received on the communication protocol. Additional we added changes to the traceback of errors on the server. We also resolved any console errors we were receiving on any of the main dashboards. Lastly we added a few checks to startup for the memoization cache, language tool, and gpt related items.