Added more distinct admin and teacher roles and ability to impersonate (#111)
Added STUDENT, TEACHER, and ADMIN roles to system. Prior we had TEACHER and ADMIN fulfilling the same role, they are now broken apart. Added capability to read in the list of approved admins and teachers.

We also updated the Google endpoint to expect a cache be setup and used. This included us adding in the filesystem cache to the list of available caches.