languagetool startup and highlight dashboard improvements
LanguageTool startup is now mostly contained within LanguageTool. Additionally, languagetool now functions with an async generator in the communication protocol.

Initially this PR was supposed to integrate LT options with the highlight dashboard but selecting errors to figure out which ones are being made makes zero sense. The options are there but not attached to the highlight dashboard as of now. We likely want some way of informing the teacher which LT items are occurring before presenting them highlight options.